python-language/image_diff/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-diff"
version = "0.1.0"
description = "A small tool to display the diff in pixels of two png files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[tool.setuptools]
py-modules = ["main", "exceptions", "image_utils"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
